+++ test/generators.test.inc.js
@@ -6,6 +6,7 @@
 // modules
 var chai = require('chai'),
 	expect = chai.expect,
+	Promise = require('bluebird'),
 	series = require('../lib/');
 
 // init