+++ changelog.md
@@ -14,6 +14,7 @@
 
 ## Next
 
+* Add `.use(Promise)` method
 * Remove redundant native-or-bluebird dependency
 * Update bluebird dependency
 * Update dev dependencies